test_opentelemetry_py/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-zipkin
opentelemetry-exporter-zipkin-json
opentelemetry-exporter-zipkin-proto-http
